Make requested changes.
